tipsum: Alter algorithm to generate true sentences.
If we find a word ending with a period (blessedly, the corpus contains no full
stops), we insert a marker into the token stream and use that later to delimit
chains. We also keep a list of words which can bootstrap a sentence, and
partial chains for them. This means that sentences can now stand alone, and
it's possible to get a good mix of both really long and really short
sentences.